+++ modules/lib/lib-export/src/main/java/com/enonic/xp/lib/export/FunctionBasedNodeImportListener.java
@@ -9,16 +9,16 @@ import static java.util.Objects.requireNonNullElse;
 public class FunctionBasedNodeImportListener
     implements NodeImportListener
 {
-    private static final Function<Long, Void> NULL_FUNCTION = aLong -> null;
+    private static final Function<Integer, Void> NULL_FUNCTION = aLong -> null;
 
-    private final Function<Long, Void> nodeImported;
+    private final Function<Integer, Void> nodeImported;
 
-    private final Function<Long, Void> nodeResolved;
+    private final Function<Integer, Void> nodeResolved;
 
-    private final Function<Long, Void> nodeSkipped;
+    private final Function<Integer, Void> nodeSkipped;
 
-    public FunctionBasedNodeImportListener( final Function<Long, Void> nodeImported, final Function<Long, Void> nodeResolved,
-                                            final Function<Long, Void> nodeSkipped )
+    public FunctionBasedNodeImportListener( final Function<Integer, Void> nodeImported, final Function<Integer, Void> nodeResolved,
+                                            final Function<Integer, Void> nodeSkipped )
     {
         this.nodeImported = requireNonNullElse( nodeImported, NULL_FUNCTION );
         this.nodeResolved = requireNonNullElse( nodeResolved, NULL_FUNCTION );
@@ -26,19 +26,19 @@ public class FunctionBasedNodeImportListener
     }
 
     @Override
-    public void nodeImported( final long count )
+    public void nodeImported( final int count )
     {
         nodeImported.apply( count );
     }
 
     @Override
-    public void nodeResolved( final long count )
+    public void nodeResolved( final int count )
     {
         nodeResolved.apply( count );
     }
 
     @Override
-    public void nodeSkipped( final long count )
+    public void nodeSkipped( final int count )
     {
         nodeSkipped.apply( count );
     }
